requirement.txt
openai==1.59.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
tavily-python==0.5.0